Route select from terminal popup, refactor stop and terminal popups.
