+++ schema.graphql
@@ -299,6 +299,7 @@ type Query {
   routes(date: Date, filter: RouteFilterInput): [Route]!
   stop(date: Date!, stopId: String!): Stop
   stops(date: Date, filter: StopFilterInput): [Stop]!
+  terminal(date: Date, terminalId: String): Terminal
   terminals(date: Date): [Terminal]!
   uiMessage: UIMessage!
   unsignedVehicleEvents(date: Date!, uniqueVehicleId: VehicleId!): [VehiclePosition]!
@@ -390,7 +391,8 @@ type Terminal implements Position {
   lng: Float!
   modes: [String!]
   name: String!
-  stops: [String!]
+  stopIds: [String!]
+  stops: [Stop!]
 }
 
 type UIMessage {
